feat(spaces): move members, invites & groups to a space Users page
Consolidates people-management for a space onto a new admin-only Users page
(/spaces/{id}/users): direct members + invite-by-email, pending invites, and
the space's groups (list + "New group", each linking to the group detail).

- General settings (/spaces/{id}/settings) drops the Members + Pending-invites
  cards; it now covers details/appearance, export, and the danger zone, and
  links over to the Users page.
- Sidebar Settings: adds an admin-only "Users" link; removes the standalone
  "Groups" link.
- The standalone /groups index is folded in — it now redirects to the active
  space's Users page. /groups/new and /groups/{id} stay.
- Drops the redundant top-level "Manage Groups" account-menu link.
- E2E groups spec rewritten for the new surfaces; CLAUDE.md updated.
